+++ configuration.php
@@ -28,13 +28,11 @@ $instance_tagline = $_POST["instancetagline"]; // This is the displayed tagline 
 $credit_level = $_POST["creditlevel"]; // This is the level of credit given to V0LT on the main page.
 $display_advanced_tools = $_POST["displayadvancedtools"]; // This determines whether or not the advanced tools section will be displayed on the tools page.
 $displayed_search_results_count = $_POST["displayedsearchresultscount"]; // This is the number of results that will be displayed in the Search tool.
+$backup_overwriting = $_POST["backupoverwriting"]; // This determines whether the database backup tool will allow the user to overwrite existing files.
 
 
-if ($display_advanced_tools == "on") {
-    $display_advanced_tools = true;
-} else {
-    $display_advanced_tools = false;
-}
+if ($display_advanced_tools == "on") { $display_advanced_tools = true; } else { $display_advanced_tools = false; } // Convert the 'display advanced tools' setting to a bool.
+if ($backup_overwriting == "on") { $backup_overwriting = true; } else { $backup_overwriting = false; } // Convert the 'backup overwriting' setting to a bool.
 
 
 if ($theme != null) { // Check to see if information was input through the form.
@@ -46,6 +44,7 @@ if ($theme != null) { // Check to see if information was input through the form.
     $config["credit_level"] = $credit_level;
     $config["display_advanced_tools"] = $display_advanced_tools;
     $config["displayed_search_results_count"] = $displayed_search_results_count;
+    $config["backup_overwriting"] = $backup_overwriting;
     file_put_contents("./configdatabase.txt", serialize($config)); // Write database changes to disk.
 }
 
@@ -95,6 +94,8 @@ if ($theme != null) { // Check to see if information was input through the form.
             <label for="displayadvancedtools">Display Advanced Tools: </label><input id="displayadvancedtools" name="displayadvancedtools" type="checkbox" <?php if ($config["display_advanced_tools"] == true) { echo "checked"; } ?>>
             <br><br>
             <label for="displayedsearchresultscount">Displayed Search Results: </label><input id="displayedsearchresultscount" name="displayedsearchresultscount" type="number" placeholder="Displayed Search Results Count" value="<?php echo $config["displayed_search_results_count"]; ?>">
+            <br><br>
+            <label for="backupoverwriting">Backup Overwriting: </label><input id="backupoverwriting" name="backupoverwriting" type="checkbox" <?php if ($config["backup_overwriting"] == true) { echo "checked"; } ?>>
             <br><br>
             <input type="submit" value="Submit">
         </form>